FIXED PARTICLE LOCATION BUG
Because Particles are children of Dot, they are drawn relative to Dot in the DisplayObjectContainer draw() method. Therefore, we don't wanna initalize Particle with the coordinates (x, y) as that will displace them another (x, y) coordinates away from the Dot. This is not intentional. We want Particles to be centered on the Dot, therefore we pass the inital coordinates (0, 0)